requirements_fleet.txt
# Fleet Simulator — Python dependencies
# Install: pip install -r requirements_fleet.txt

grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
pyyaml>=6.0
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
cryptography>=42.0
